refactor(property-vals): inline AppContext into Config
AppContext was a 1:1 wrapper over five Config fields plus a
should_process method. Workers can read Config directly; should_process
moves to Config alongside the fields it reads, with the filter logic
extracted to a free function so tests can exercise it without
constructing a Config.
