+++ rust/property-vals-rs/src/config.rs
@@ -1,8 +1,10 @@
+use std::hash::{Hash, Hasher};
 use std::{num::ParseIntError, str::FromStr};
 
 use common_continuous_profiling::ContinuousProfilingConfig;
 use common_kafka::config::{ConsumerConfig, KafkaConfig};
 use envconfig::Envconfig;
+use siphasher::sip::SipHasher13;
 
 #[derive(Envconfig, Clone)]
 pub struct Config {
@@ -15,65 +17,36 @@ pub struct Config {
     #[envconfig(nested = true)]
     pub consumer: ConsumerConfig,
 
-    /// Topic the aggregator emits flush messages to. CH's `kafka_property_values`
-    /// engine table consumes from this and writes to `posthog.property_values`.
     #[envconfig(default = "clickhouse_property_values")]
     pub output_topic: String,
 
-    /// How long to accumulate aggregator state before flushing. Each flush
-    /// emits one message per unique tuple seen during the window.
     #[envconfig(default = "30")]
     pub flush_interval_secs: u64,
 
-    /// Cap on the worker's in-memory tuple buffer. Hitting this triggers
-    /// an early backpressure flush ahead of the timer.
     #[envconfig(default = "500000")]
     pub max_buffered_tuples: usize,
 
-    /// Transactional producer ID for the events worker. Must be unique per
-    /// pod and stable across restarts (e.g. K8s pod name + "-events"). The
-    /// groups worker uses a separate id because rdkafka allows only one
-    /// outstanding transaction per `transactional.id`.
     #[envconfig(default = "property-vals-rs-local-events")]
     pub kafka_transactional_id: String,
 
-    /// How long Kafka will wait on init_transactions, send_offsets_to_transaction,
-    /// and commit_transaction calls before timing out the transaction.
     #[envconfig(default = "60")]
     pub kafka_transaction_timeout_secs: u64,
 
-    /// Topic that carries `$groupidentify` messages. Each message is one
-    /// group update with the full property blob.
     #[envconfig(default = "clickhouse_groups")]
     pub groups_kafka_consumer_topic: String,
 
-    /// Consumer group for the groups worker. Independent of the events
-    /// consumer group; the two workers are unrelated for Kafka's purposes.
     #[envconfig(default = "clickhouse-property-vals-rs-groups")]
     pub groups_kafka_consumer_group: String,
 
-    /// Transactional producer ID for the groups worker. Distinct from
-    /// `kafka_transactional_id` because each producer needs its own id.
     #[envconfig(default = "property-vals-rs-local-groups")]
     pub groups_kafka_transactional_id: String,
 
-    /// Teams that are always processed regardless of `rollout_percentage`.
-    /// Use to pin specific teams during testing or keep a hand-picked
-    /// team included while ramping the rest of the fleet.
     #[envconfig(default = "")]
     pub allowed_teams: TeamList,
 
-    /// Teams that are never processed regardless of allow-list or rollout.
-    /// `blocked_teams` wins over `allowed_teams` on overlap, since "drop
-    /// this team" is a safer default than "process this team".
     #[envconfig(default = "")]
     pub blocked_teams: TeamList,
 
-    /// Percentage of teams (0-100) to process, picked by a stable hash
-    /// of `team_id`. `allowed_teams` overrides this; `blocked_teams` still
-    /// excludes. Default in code is 100 (process all) so `cargo run`
-    /// against the local stack does something visible. Production charts
-    /// set this to 0 and ramp via overrides.
     #[envconfig(default = "100")]
     pub rollout_percentage: u8,
 
@@ -106,14 +79,9 @@ impl FromStr for TeamList {
 
 impl Config {
     pub fn init_with_defaults() -> Result<Self, envconfig::Error> {
-        // Default to clickhouse_events_json so cargo run works against the local
-        // dev stack with no env overrides. Production charts set
-        // KAFKA_CONSUMER_TOPIC=team_event_partitioned_events_json.
-        //
         // auto_commit=false because offsets are committed by the
-        // transactional producer via send_offsets_to_transaction. If
-        // librdkafka auto-committed in the background we'd lose the
-        // exactly-once guarantee.
+        // transactional producer via send_offsets_to_transaction; auto-commit
+        // would break the exactly-once guarantee.
         ConsumerConfig::set_defaults(
             "clickhouse-property-vals-rs",
             "clickhouse_events_json",
@@ -121,11 +89,48 @@ impl Config {
         );
         Config::init_from_env()
     }
+
+    pub fn should_process(&self, team_id: i64) -> bool {
+        team_filter_allows(
+            team_id,
+            &self.allowed_teams.teams,
+            &self.blocked_teams.teams,
+            self.rollout_percentage,
+        )
+    }
+}
+
+pub fn team_filter_allows(
+    team_id: i64,
+    allowed: &[i64],
+    blocked: &[i64],
+    rollout_percentage: u8,
+) -> bool {
+    if blocked.contains(&team_id) {
+        return false;
+    }
+    if allowed.contains(&team_id) {
+        return true;
+    }
+    team_in_rollout(team_id, rollout_percentage)
+}
+
+fn team_in_rollout(team_id: i64, rollout_percentage: u8) -> bool {
+    if rollout_percentage >= 100 {
+        return true;
+    }
+    if rollout_percentage == 0 {
+        return false;
+    }
+    let mut hasher = SipHasher13::new();
+    team_id.hash(&mut hasher);
+    (hasher.finish() % 100) < rollout_percentage as u64
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use proptest::prelude::*;
 
     #[test]
     fn team_list_parses() {
@@ -133,5 +138,79 @@ mod tests {
         assert_eq!(list.teams, vec![1, 2, 3]);
         let empty: TeamList = "".parse().unwrap();
         assert!(empty.teams.is_empty());
+    }
+
+    fn arb_team_list() -> impl Strategy<Value = Vec<i64>> {
+        prop::collection::vec(-1_000i64..=1_000, 0..10)
+    }
+
+    proptest! {
+        #[test]
+        fn blocked_teams_never_process(
+            team_id: i64,
+            rollout in 0u8..=100,
+            allowed in arb_team_list(),
+            mut blocked in arb_team_list(),
+        ) {
+            blocked.push(team_id);
+            prop_assert!(!team_filter_allows(team_id, &allowed, &blocked, rollout));
+        }
+
+        #[test]
+        fn allowed_teams_process_when_not_blocked(
+            team_id: i64,
+            rollout in 0u8..=100,
+            blocked in arb_team_list(),
+            mut allowed in arb_team_list(),
+        ) {
+            let blocked: Vec<i64> = blocked.into_iter().filter(|t| *t != team_id).collect();
+            allowed.push(team_id);
+            prop_assert!(team_filter_allows(team_id, &allowed, &blocked, rollout));
+        }
+
+        #[test]
+        fn full_rollout_processes_any_non_blocked_team(
+            team_id: i64,
+            allowed in arb_team_list(),
+            blocked in arb_team_list(),
+        ) {
+            let blocked: Vec<i64> = blocked.into_iter().filter(|t| *t != team_id).collect();
+            prop_assert!(team_filter_allows(team_id, &allowed, &blocked, 100));
+        }
+
+        #[test]
+        fn zero_rollout_drops_any_non_allowed_team(
+            team_id: i64,
+            allowed in arb_team_list(),
+            blocked in arb_team_list(),
+        ) {
+            let allowed: Vec<i64> = allowed.into_iter().filter(|t| *t != team_id).collect();
+            let blocked: Vec<i64> = blocked.into_iter().filter(|t| *t != team_id).collect();
+            prop_assert!(!team_filter_allows(team_id, &allowed, &blocked, 0));
+        }
+
+        #[test]
+        fn should_process_is_deterministic(
+            team_id: i64,
+            rollout in 0u8..=100,
+            allowed in arb_team_list(),
+            blocked in arb_team_list(),
+        ) {
+            let first = team_filter_allows(team_id, &allowed, &blocked, rollout);
+            for _ in 0..5 {
+                prop_assert_eq!(team_filter_allows(team_id, &allowed, &blocked, rollout), first);
+            }
+        }
+    }
+
+    #[test]
+    fn rollout_percentage_approximates_target_share() {
+        let included = (1..=10_000)
+            .filter(|t| team_filter_allows(*t, &[], &[], 10))
+            .count();
+        assert!(
+            (900..=1100).contains(&included),
+            "expected ~1000 of 10000 at 10%, got {included}"
+        );
     }
 }
