+++ rust/property-vals-rs/src/lib.rs
@@ -1,2 +1,1 @@
-pub mod app_context;
 pub mod config;
